Relocate source files and update build scripts and versions
The source files have been moved from the 'src' folder to a more appropriately named 'load-html-files' folder. This led to updates in the Gulp build and translation scripts, phpcs rule files and the addition of wp-env.json for easier local development setup. Additionally, outdated dependencies in the composer.json and copyright dates have been updated to the current state.