Add 'wp_autoloader' library and update dependencies
The commit includes the addition of the 'wp_autoloader' library to the project dependencies. It also includes a change in the way the AutoloaderPlugin is utilized in 'load-html-files.php'. The 'phpstan' dependency version was updated as well in 'composer.lock', and corresponding content-hash value and reference details were adjusted accordingly.